# Conflicts:
#	chain/rpc/src/client.rs
#	hopr/hopr-lib/tests/chain_integration_tests.rs

+++ db/entity/src/conversions/tickets.rs
@@ -1,9 +1,11 @@
-use crate::errors::DbEntityError;
-use crate::ticket;
+use sea_orm::Set;
+
 use hopr_crypto_types::prelude::*;
 use hopr_internal_types::prelude::*;
 use hopr_primitive_types::prelude::*;
-use sea_orm::Set;
+
+use crate::errors::DbEntityError;
+use crate::ticket;
 
 impl TryFrom<&ticket::Model> for AcknowledgedTicket {
     type Error = DbEntityError;
@@ -68,7 +70,7 @@ impl From<AcknowledgedTicket> for ticket::ActiveModel {
             channel_epoch: Set(U256::from(value.verified_ticket().channel_epoch).to_be_bytes().to_vec()),
             signature: Set(value.verified_ticket().signature.unwrap().as_ref().to_vec()),
             response: Set(value.response.as_ref().to_vec()),
-            state: Set(value.status as u8 as i32),
+            state: Set(value.status as i8),
             hash: Set(value.ticket.verified_hash().as_ref().to_vec()),
             ..Default::default()
         }
